// frontend/src/generated/types/dailySale.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * QROrder API
 * QROrder Management System API 문서
 * OpenAPI spec version: v1.0
 */

export interface DailySale {
  groupDate?: string;
  dayTotalPrice?: number;
  dayCancelPrice?: number;
  dayNetPrice?: number;
  dayOrderCount?: number;
  dayCancelCount?: number;
}