// frontend/src/generated/types/settlementResponse.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * QROrder API
 * QROrder Management System API 문서
 * OpenAPI spec version: v1.0
 */
import type { DailySale } from './dailySale';

export interface SettlementResponse {
  totalPrice?: number;
  cancelPrice?: number;
  discountPice?: number;
  netPrice?: number;
  orderCount?: number;
  dailySales?: DailySale[];
}